+++ include/tools/confbuild.hpp
@@ -15,9 +15,10 @@
 // [2021y-03m-09d][01:44:12] Idrisov Denis R. 016
 // [2021y-03m-09d][21:02:02] Idrisov Denis R. 017
 // [2021y-03m-11d][00:30:56] Idrisov Denis R. 018
+// [2021y-03m-11d][12:43176] Idrisov Denis R. 019 PRE
 #pragma once
 #ifndef dTOOLS_CONFBUILD_USED_
-#define dTOOLS_CONFBUILD_USED_ 018
+#define dTOOLS_CONFBUILD_USED_ 019 PRE
 //==============================================================================
 //==============================================================================
 
@@ -57,7 +58,7 @@
 
 #define dVERSION_MAJOR      0
 #define dVERSION_MINOR      1
-#define dVERSION_PATCH      8
+#define dVERSION_PATCH      9
 
 #define dSTRINGIZE(...) #__VA_ARGS__
 #define dSSTRINGIZE(x) dSTRINGIZE(x)