+++ test/test-200-types/sources/tools/types/test-fixed.cpp
@@ -144,7 +144,6 @@ TEST_COMPONENT(009)
     );
 }
 
-
 //==============================================================================
 //=== macro ====================================================================
 
@@ -162,15 +161,23 @@ TEST_COMPONENT(010)
     me::int64_t min64 = INT64_MIN;
     me::int64_t max64 = INT64_MAX;
 
-    dSTATIC_ASSERT(me::map_signed<1>::min_value == INT8_MIN, ERROR_INT8_MIN );
-    dSTATIC_ASSERT(me::map_signed<2>::min_value == INT16_MIN, ERROR_INT16_MIN);    
-    dSTATIC_ASSERT(me::map_signed<4>::min_value == INT32_MIN, ERROR_INT32_MIN);
-    dSTATIC_ASSERT(me::map_signed<8>::min_value == INT64_MIN, ERROR_INT64_MIN);
+    dSTATIC_ASSERT(me::map_signed<1>::min_value == INT8_MIN, 
+        ERROR_INT8_MIN );
+    dSTATIC_ASSERT(me::map_signed<2>::min_value == INT16_MIN, 
+        ERROR_INT16_MIN);    
+    dSTATIC_ASSERT(me::map_signed<4>::min_value == INT32_MIN, 
+        ERROR_INT32_MIN);
+    dSTATIC_ASSERT(me::map_signed<8>::min_value == INT64_MIN,
+        ERROR_INT64_MIN);
 
-    dSTATIC_ASSERT(me::map_signed<1>::max_value == INT8_MAX, ERROR_INT8_MAX );
-    dSTATIC_ASSERT(me::map_signed<2>::max_value == INT16_MAX, ERROR_INT16_MAX);    
-    dSTATIC_ASSERT(me::map_signed<4>::max_value == INT32_MAX, ERROR_INT32_MAX);
-    dSTATIC_ASSERT(me::map_signed<8>::min_value == INT64_MIN, ERROR_INT64_MAX);
+    dSTATIC_ASSERT(me::map_signed<1>::max_value == INT8_MAX, 
+        ERROR_INT8_MAX );
+    dSTATIC_ASSERT(me::map_signed<2>::max_value == INT16_MAX, 
+        ERROR_INT16_MAX);    
+    dSTATIC_ASSERT(me::map_signed<4>::max_value == INT32_MAX, 
+        ERROR_INT32_MAX);
+    dSTATIC_ASSERT(me::map_signed<8>::min_value == INT64_MIN,
+        ERROR_INT64_MAX);
 
     ASSERT_TRUE(min8  == me::map_signed<1>::min_value);
     ASSERT_TRUE(min16 == me::map_signed<2>::min_value);
@@ -197,15 +204,23 @@ TEST_COMPONENT(011)
     me::uint64_t min64 = 0;
     me::uint64_t max64 = UINT64_MAX;
 
-    dSTATIC_ASSERT(me::map_unsigned<1>::min_value == 0, ERROR_UINT8_MIN );
-    dSTATIC_ASSERT(me::map_unsigned<2>::min_value == 0, ERROR_UINT16_MIN);    
-    dSTATIC_ASSERT(me::map_unsigned<4>::min_value == 0, ERROR_UINT32_MIN);
-    dSTATIC_ASSERT(me::map_unsigned<8>::min_value == 0, ERROR_UINT64_MIN);
+    dSTATIC_ASSERT(me::map_unsigned<1>::min_value == 0, 
+        ERROR_UINT8_MIN );
+    dSTATIC_ASSERT(me::map_unsigned<2>::min_value == 0, 
+        ERROR_UINT16_MIN);    
+    dSTATIC_ASSERT(me::map_unsigned<4>::min_value == 0, 
+        ERROR_UINT32_MIN);
+    dSTATIC_ASSERT(me::map_unsigned<8>::min_value == 0, 
+        ERROR_UINT64_MIN);
 
-    dSTATIC_ASSERT(me::map_unsigned<1>::max_value == UINT8_MAX, ERROR_UINT8_MAX );
-    dSTATIC_ASSERT(me::map_unsigned<2>::max_value == UINT16_MAX, ERROR_UINT16_MAX);    
-    dSTATIC_ASSERT(me::map_unsigned<4>::max_value == UINT32_MAX, ERROR_UINT32_MAX);
-    dSTATIC_ASSERT(me::map_unsigned<8>::max_value == UINT64_MAX, ERROR_UINT64_MAX);
+    dSTATIC_ASSERT(me::map_unsigned<1>::max_value == UINT8_MAX,
+        ERROR_UINT8_MAX );
+    dSTATIC_ASSERT(me::map_unsigned<2>::max_value == UINT16_MAX, 
+        ERROR_UINT16_MAX);    
+    dSTATIC_ASSERT(me::map_unsigned<4>::max_value == UINT32_MAX,
+        ERROR_UINT32_MAX);
+    dSTATIC_ASSERT(me::map_unsigned<8>::max_value == UINT64_MAX,
+        ERROR_UINT64_MAX);
 
     ASSERT_TRUE(min8  == me::map_unsigned<1>::min_value);
     ASSERT_TRUE(min16 == me::map_unsigned<2>::min_value);